Safari-609.1.11.1/LayoutTests/http/tests/quicklook/same-origin-xmlhttprequest-allowed.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/ui-helper.js"></script>
<script src="resources/tap-run-test-hyperlink.js"></script>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFramesAsText();
    testRunner.waitUntilDone();
}
</script>
</head>
<body>
<iframe src="resources/same-origin-xmlhttprequest-allowed.rtf" onload="runTest(this, 8 /* offsetInFrame */, true /* runAsync */)"></iframe>
</body>
</html>
