LayoutTests/http/tests/quicklook/same-origin-xmlhttprequest-allowed-expected.txt

CONSOLE MESSAGE: line 1: PASS: XMLHttpRequest allowed


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Run test
